+++ client/src/components/EventCard/EventCard.tsx
@@ -1,55 +1,26 @@
-import { Card, Typography, Chip, IconButton, Box, styled, Theme, SxProps, Dialog, Button, Menu, MenuItem, Slide, AppBar } from '@mui/material';
-import HourglassBottomRoundedIcon from '@mui/icons-material/HourglassBottomRounded';
-import MeetingRoomRoundedIcon from '@mui/icons-material/MeetingRoomRounded';
+import { Typography, Chip, IconButton, Box, styled, Theme, SxProps, Menu, MenuItem } from '@mui/material';
 import InsertLinkRoundedIcon from '@mui/icons-material/InsertLinkRounded';
-import StairsIcon from '@mui/icons-material/Stairs';
 import React, { useEffect, useState } from 'react';
-import AccessTimeFilledRoundedIcon from '@mui/icons-material/AccessTimeFilledRounded';
-import ArrowBackIosRoundedIcon from '@mui/icons-material/ArrowBackIosRounded';
-import PeopleRoundedIcon from '@mui/icons-material/PeopleRounded';
 import FiberManualRecordIcon from '@mui/icons-material/FiberManualRecord';
-import { convertToLocaleTime, createDropdownOptions, populateDurationOptions } from '../helpers/utility';
-import { FormData } from '../helpers/types';
 import toast from 'react-hot-toast';
 import { useNavigate } from 'react-router-dom';
-import { ROUTES } from '../config/routes';
-import Api from '../api/api';
 import { EventResponse } from '@bookify/shared';
 import MoreVertIcon from '@mui/icons-material/MoreVert';
-import { TransitionProps } from '@mui/material/transitions';
-import Dropdown, { DropdownOption } from './Dropdown';
-
-interface ChipData {
-  icon: React.ReactElement;
-  label: string;
-  color?: string;
-  type?: 'conference' | 'floor' | 'seats' | 'time' | 'room';
-}
-
-interface EventCardProps {
-  sx?: SxProps<Theme>;
-  event?: EventResponse;
-  onDelete: (id?: string) => void;
-  disabled?: boolean;
-  onEdit: (id: string, data: any) => void;
-}
-
-type EditRoomFields = Pick<FormData, 'duration'>;
-
-const Transition = React.forwardRef(function Transition(
-  props: TransitionProps & {
-    children: React.ReactElement<any>;
-  },
-  ref: React.Ref<unknown>,
-) {
-  return <Slide direction="left" ref={ref} {...props} />;
-});
+import EditDialog from './EditDialog';
+import MeetingRoomRoundedIcon from '@mui/icons-material/MeetingRoomRounded';
+import StairsIcon from '@mui/icons-material/Stairs';
+import AccessTimeFilledRoundedIcon from '@mui/icons-material/AccessTimeFilledRounded';
+import PeopleRoundedIcon from '@mui/icons-material/PeopleRounded';
+import { EditRoomFields } from './util';
+import { ROUTES } from '../../config/routes';
+import Api from '../../api/api';
+import { convertToLocaleTime } from '../../helpers/utility';
 
 const ListItem = styled('li')(({ theme }) => ({
   margin: theme.spacing(0.3),
 }));
 
-const createChips = (event: EventResponse) => {
+export const createChips = (event: EventResponse) => {
   return [
     {
       label: convertToLocaleTime(event?.start) + ' - ' + convertToLocaleTime(event?.end),
@@ -71,27 +42,34 @@ const createChips = (event: EventResponse) => {
   ];
 };
 
+interface EventCardProps {
+  sx?: SxProps<Theme>;
+  event?: EventResponse;
+  onDelete: (id?: string) => void;
+  disabled?: boolean;
+  onEdit: (id: string, data: any) => void;
+}
+
+interface ChipData {
+  icon: React.ReactElement;
+  label: string;
+  color?: string;
+  type?: 'conference' | 'floor' | 'seats' | 'time' | 'room';
+}
+
 const EventCard = ({ sx, event, onDelete, disabled, onEdit }: EventCardProps) => {
   const [chips, setChips] = useState<ChipData[]>([]);
   const [isOngoingEvent, setIsOngoingEvent] = useState(false);
   const [editDialogOpen, setEditDialogOpen] = useState(false);
+
   const [loading, setLoading] = useState(false);
   const navigate = useNavigate();
   const [formData, setFormData] = useState<EditRoomFields>({
     duration: 30,
   });
-  const [durationOptions, setDurationOptions] = useState<DropdownOption[]>([]);
 
   const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
   const open = Boolean(anchorEl);
-
-  // TODO: create separate component for the dialog
-  useEffect(() => {
-    if (editDialogOpen) {
-      const durationOptions = populateDurationOptions(30, 3 * 60); // 30 mins -> 5 hrs
-      setDurationOptions(createDropdownOptions(durationOptions, 'time'));
-    }
-  }, [editDialogOpen]);
 
   useEffect(() => {
     if (event) {
@@ -150,7 +128,6 @@ const EventCard = ({ sx, event, onDelete, disabled, onEdit }: EventCardProps) =>
     if (res?.status === 'error') {
       res.message && toast.error(res.message);
       setLoading(false);
-      setEditDialogOpen(false);
       return;
     }
 
@@ -278,119 +255,7 @@ const EventCard = ({ sx, event, onDelete, disabled, onEdit }: EventCardProps) =>
         })}
       </Box>
 
-      {/* edit dialog*/}
-      <Dialog fullScreen open={editDialogOpen} onClose={() => setEditDialogOpen(false)} TransitionComponent={Transition}>
-        <AppBar
-          sx={{ bgcolor: 'transparent', position: 'relative', display: 'flex', flexDirection: 'row', py: 2, alignItems: 'center', px: 4, boxShadow: 'none' }}
-        >
-          <IconButton edge="start" color="inherit" onClick={() => setEditDialogOpen(false)} aria-label="close">
-            <ArrowBackIosRoundedIcon
-              fontSize="small"
-              sx={[
-                (theme) => ({
-                  color: theme.palette.common.black,
-                }),
-              ]}
-            />
-          </IconButton>
-          <Typography
-            sx={[
-              (theme) => ({
-                textAlign: 'center',
-                flex: 1,
-                color: theme.palette.common.black,
-                fontWeight: 700,
-              }),
-            ]}
-            variant="h5"
-            component={'div'}
-          >
-            Edit event
-          </Typography>
-        </AppBar>
-
-        <Box
-          mt={2}
-          mx={4}
-          sx={{
-            py: 1,
-            px: 2,
-            bgcolor: 'white',
-            borderTopLeftRadius: 10,
-            borderTopRightRadius: 10,
-            borderBottomLeftRadius: 15,
-            borderBottomRightRadius: 15,
-          }}
-        >
-          <Dropdown
-            id="duration"
-            options={durationOptions}
-            value={formData.duration.toString()}
-            onChange={handleInputChange}
-            icon={
-              <HourglassBottomRoundedIcon
-                sx={[
-                  (theme) => ({
-                    color: theme.palette.grey[50],
-                  }),
-                ]}
-              />
-            }
-          />
-        </Box>
-
-        <Box flexGrow={1} />
-        <Box
-          sx={{
-            mx: 4,
-            mb: 3,
-            textAlign: 'center',
-          }}
-        >
-          <Button
-            onClick={onEditRoomClick}
-            fullWidth
-            variant="contained"
-            disableElevation
-            sx={[
-              (theme) => ({
-                py: 2,
-                backgroundColor: theme.palette.common.white,
-                borderRadius: 15,
-
-                color: theme.palette.common.black,
-              }),
-            ]}
-          >
-            <Typography variant="h6" fontWeight={700} color="error">
-              Save
-            </Typography>
-          </Button>
-          <Button
-            variant="text"
-            onClick={() => setEditDialogOpen(false)}
-            sx={{
-              py: 2,
-              mt: 2,
-              px: 3,
-              boxShadow: 'none',
-              '&:hover': {
-                boxShadow: 'none',
-              },
-              '&:active': {
-                boxShadow: 'none',
-              },
-              '&:focus': {
-                boxShadow: 'none',
-              },
-            }}
-          >
-            <Typography variant="subtitle2" fontWeight={700}>
-              Dismiss
-            </Typography>
-          </Button>
-        </Box>
-      </Dialog>
+      <EditDialog open={editDialogOpen} setOpen={setEditDialogOpen} onChange={handleInputChange} data={formData} onEditRoomClick={onEditRoomClick} />
     </Box>
   );
 };
