+++ client/src/components/EventCard/index.tsx
@@ -1,0 +1,3 @@
+import EventCard from './EventCard';
+
+export { EventCard };
