+++ client/src/components/EventCard/util.ts
@@ -1,0 +1,3 @@
+import { FormData } from '../../helpers/types';
+
+export type EditRoomFields = Pick<FormData, 'duration'>;
