+++ client/src/helpers/types.ts
@@ -2,7 +2,6 @@ export interface FormData {
   startTime: string;
   duration: number;
   seats: number;
-  floor: string;
   title?: string;
   attendees?: string[];
   conference?: boolean;
