+++ client/src/pages/Home/AdvancedOptionsDialog.tsx
@@ -2,10 +2,8 @@ import * as React from 'react';
 import Button from '@mui/material/Button';
 import Dialog from '@mui/material/Dialog';
 import AppBar from '@mui/material/AppBar';
-import Toolbar from '@mui/material/Toolbar';
 import IconButton from '@mui/material/IconButton';
 import Typography from '@mui/material/Typography';
-import CloseIcon from '@mui/icons-material/Close';
 import Slide from '@mui/material/Slide';
 import { TransitionProps } from '@mui/material/transitions';
 import { Box, Checkbox, Divider } from '@mui/material';
@@ -58,7 +56,7 @@ export default function AdvancedOptionsDialog({ handleInputChange, open, handleC
           variant="h5"
           component={'div'}
         >
-          Advance options
+          Additional options
         </Typography>
       </AppBar>
       <Box
