+++ client/src/pages/Home/shared.ts
@@ -1,0 +1,14 @@
+import { DropdownOption } from '../../components/Dropdown';
+
+export interface Event {
+  room?: string;
+  eventId?: string;
+  start?: string;
+  end?: string;
+  summary?: string;
+  availableRooms?: DropdownOption[];
+  roomEmail?: string;
+  seats?: number;
+  isEditable?: boolean;
+  createdAt?: number;
+}
